parser, eval: use an ut_op flag to denote postfix access
The current code still abused the JSON value pointer to denote postfix
access for certain operations which led to a crash when freeing the
parser state due to an attempt to put a (void *)1 pointer.

Since we do have the ability to set flags on operations since the AST
rework, use this much cleaner approach and avoid the invalid pointer
hackery.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>