eval: restore context pointer of first evaluated dot/bracket expression
When an expression such as `foo.bar(a.b, c.d)` is evaluated, the state
context pointer will point to `c` while we need `foo` when invoking
functions.

Since the context pointer is only relevant for function calls and since
for function call opcodes, the lhs expression is always the first operand,
there is no need to store the context of subsequent ops.

Adjust the ut_get_operands() procedure to restore state->ctx to the result
of the first evaluated operand.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>