eval: fix potential segfault in for() loops
When a for() loop body, initializer, test or increment expression compiles
further code, e.g. by invoking invoke() or require(), the opcode pool will
be reallocated, potentially changing the addresses of all opcodes.

This might lead to an invalid memory access when previously cached opcode
pointers are accessed later on.

Solve this issue by obtaining the relative offsets of the corresponding
opcodes, avoiding the need for pointer dereferences.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>