lib: retain prototype when passing scopes to include()
Change uc_include() to retain the prototype of the given scope object when
processing includes. Also change the default behaviour to register the
current VM scope as prototype on the passed scope object so that included
code has access to functions such as length(), print() etc. by default.

To actually sandbox the included code, the new `proto()` function can be
used to create a scope object with an empty prototype:

  `include(..., proto({ ... }, {}))`

Signed-off-by: Jo-Philipp Wich <jo@mein.io>