lib: implement `sourcepath()` function
The sourcepath() function allows querying the filesystem path of the source
file currently being executed by ucode.

The optional depth argument can be used to walk up the include stack to
determine the path of the file that included the current file, the path of
the parent file of the parent file and so on.

By specifying a truish value as second argument, only the directory portion
of the source file path is returned. This is useful to e.g. discover
ressources relative to the current source file directory.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>