lib, vm: ensure that require() compiles modules only once
Cache the result of a successful require operation in a global.modules
object and return the cached value for subsequent require calls on the
same module name.

This ensures that a given module is only compiled and executed once,
regardless of how many times it is used.

A reload of the module can be forced by deleting the corresponding
key in the global module table.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>