vm: fix invalid memory access on toplevel function calls
When a function is invoked directly, e.g. through uc_vm_invoke(), the call
stack is empty, so avoid accessing the first call frame unless we actually
need to, which is only the case if the function is invoked with spread args
which can only happen with calls from manged code.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>