vm: make root exception handler configurable
So far, the VM simply printed exception information to stderr if the
exception was not catched in managed code. Host programs embedding
ucode might want to customize that behaviour, so refactor the current
defualt behaviour into a callback function and add a public getter
and setter to allow changing the exception handler callback.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>