main: refactor option parsing and VM setup
Set VM options and environment variables and load modules on the fly
while parsing the cli options instead of aggregating all the values in
memory first.

This vastly reduces the amount of arguments we need to pass to the
parse() function. Also rename parse() to compile() while we're at it.

Also slightly adjust the usage output.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>