vm: fix invalid memory access on GC'ing uninitialized VM context
When attempting to invoke uc_vm_free() or uc_gc() on a uc_vm_t struct
that has not been initialized by uc_vm_init(), the circular double
linked value list is not set up, causing the GC to read invalid memory
locations when attempting to traverse the object list.

Back out early when the list heads are not properly set up in order to
prevent this issue.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>