vm: clear exception information before calling managed code functions
If execution in an existing VM that threw an exception was resumed
through uc_vm_call() or uc_vm_invoke(), the exception was never cleared,
causing all subsequent calls to return with an exception status as well.

Ensure that any preexisting exception information is discarded before
executing the requested function in order to start from a clean state.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>