fs: add utility functions
Add three new functions `dirname()`, `basename()` and `lsdir()`.

The `basename()` and `dirname()` functions behave like their libc
counterparts and return the filename and directory portion of a
given path respectively. If the path argument is missing or not
a string, null is returned.

Examples:

  dirname("/usr/lib")  -> "/usr"
  dirname("/usr/")     -> "/"
  dirname("usr")       -> "."
  dirname("/")         -> "/"
  dirname(".")         -> "."
  dirname("..")        -> "."

  basename("/usr/lib") -> "lib"
  basename("/usr/")    -> "usr"
  basename("usr")      -> "usr"
  basename("/")        -> "/"
  basename(".")        -> "."
  basename("..")       -> ".."

The `lsdir()` function returns a sorted array containing the names
of all entries within the given directory path, without the common
"." and ".." entries. If the given path is not a directory, cannot
be opened or if another system level occurs, null is returned and
`fs.error()` can be used to query details.

The function takes an optional second argument which may be either a
regular expression value or a string. In case a regular expression
is given, each directory entry is matched against it. In case a
string is provided, it is treated as wildcard (glob) pattern and
only directory entries matching the pattern are considered.

Examples:

  lsdir("/sys/class/net")            -> [ "eth0", "lo", "wlan0" ]
  lsdir("/proc", /^[0-9]+$/)         -> [ "1", "4", "12", ... ]
  lsdir("/sys/class/block/", "sd?3") -> [ "sde3", "sdf3" ]

Signed-off-by: Jo-Philipp Wich <jo@mein.io>