vm: introduce value registry
Introduce a new, lazily allocated value registry which can be used by C
code to store values which should not be garbage collected.

The registry is a plain ucode object internally and treated as GC root
but not exposed to ucode script code, this allows it to retain references
to values which are otherwise completely unreachable from ucode scripts.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>