vm: optimize string concatenation
When concatenating strings, avoid allocating three times the required
memory in the worst case. Instead of first allocating the string
representations of the operands followed by the memory for the final
string, allocate a string buffer and print the operands into it. This
will grow the target memory as needed and avoid redundant internal
copies of the involved strings.

Also handle the special where the final string fits into a tagged
pointer and deal with it accordingly.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>