lib: implement uniq() function
The uniq() function allows extracting all unique values from a given input
array in an efficient manner. It is roughly equivalent to the following
ucode idiom:

    let seen = {}
    let unique = filter(array, item => !seen[item]++);

In contrast to the code above, `uniq()` does not rely on implicit
stringification of item values but performs strict equality tests internally.

If equivalence of stringified results is desired, the following code can
be used:

    let unique = uniq(map(array, item => "" + item));

Signed-off-by: Jo-Philipp Wich <jo@mein.io>