vm: fix leaking function value on call exception
The internal uc_vm_call_function() helper may fail in different ways before
the stack frame has been set up, e.g. if the provided function value was not
actually a callable function. In such cases an exception is raised but the
actual function value is leaked since there's not yet a stackframe referring
to it.

Solve the issue by freeing the function value explicitly in these exit cases.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>