lib: refactor `uc_int()`
For string cases, turn `int()` into a thin `strtoll()` wrapper which
attempts to parse the initial portion of the string as a decimal integer
literal, optionally preceded by white space and a sign character.

Also introduce an optional `base` argument for string cases while we're
at it and adjust the existing stdlib test case accordingly.

The function now behaves mostly the same as ECMAScript `parseInt(val, 10)`
for string cases, means it will recognize `012` as `12` and not `10` and
it will accept trailing non-digit characters after the initial portition
of the input string.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>