rtnl: fix premature netlink reply receive abort
The nl_recvmsgs() logic in uc_nl_request() incorrectly stopped reading
the socket before the netlink ACK message was handled for non-multipart
replies.

This caused subsequent requests to incorrectly receive the ACK of the
previous request, leading to a failure to receive the actual reply.

Fix this issue by continue reading the socket until either the finish
callback for multipart (dump) messages or the ack callback for non-
multipart messages was received.

This fix is basically the same as the one applied to the nl80211 module
in 54ef6c0 ("nl80211: fix premature netlink reply receive abort").

Signed-off-by: Jo-Philipp Wich <jo@mein.io>