rtnl: update NETLINK_GET_STRICT_CHK socket flag with every request
So far the NETLINK_GET_STRICT_CHK socket flag was only set on the implicit
socket creation performed during the first request and ignored for
subsequent ones which made it impossible to perform only some requests
with enabled strict checking.

Modify the logic to check the flag state for every request and change
it if needed. This allows performing both strict and non-strict requests
over the same connection.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>