rtnl: fix parsing/creation of IFLA_AF_SPEC RTA for the AF_BRIDGE family
Some pecularities in the encoding of the IFLA_AF_SPEC attribute make it
unsuitable for table driven parsing/generation.

To solve this issue, introduce specific datatype handling for IFLA_AF_SPEC
and parse/generate the RTA depending on the address family of the containing
netlink message.

Also add some missing constants while we're at it.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>