compiler: resolve predeclared upvalues
Do not require a parent function compiler reference to lookup an already
declared (potentially unresolved) upvalue in the current scope. Instead,
search the named upvalues in the current function scope in case there is
no parent compiler reference.

This is required for the upcoming module support which will use unresolved
upvalues to realize import/export functionality.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>