vm: introduce import and export opcodes
Introduce new opcodes to realize module imports and exports. The export
operation will capture a local variable as upvalue and store it in VM
wide module export registry while the import operation will connect an
upvalue from the module export registry with a preallocated upvalue in
the running function scope.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>