uloop: task: gracefully handle absent output callback
Both input and output callbacks for uloop tasks are optional, but the
low level io callback implementation did not properly deal with an absent
ucode output callback, triggering an exception in managed code due to
invoking a null value as function.

Fix this issue by checking for the availability of a callable output
function and simply discarding the received task message otherwise.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>