uci: fix ctx.commit() without arguments
A uci commit operation may invalidate the package pointer, leading to an
infinite loop while iterating the packages to commit.

Refactor the code to first build a string list of configurations, then
iterating it in order to avoid the iterator invalidation.

While we're at it, also allow restricting save and revert operations to
single configs, which was rejected as invalid before.

Fixes: #146
Signed-off-by: Jo-Philipp Wich <jo@mein.io>