uloop: interrupt on VM signals
When the VM instance loading the uloop module has signal dispatching
enabled, implicitly register the signal pipe as file descriptor in the
global uloop and dispatch received signals to the VM instance.

This ensures that the respective managed ucode signal handlers are
invoked immediately and not just after `uloop_run()` returns.

Also end the uloop in case any invoked signal handler threw an
exception, to match the expected behaviour with other kinds of
callbacks.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>