vscode: implement a language server extension for Visual Studio Code
This commit introduces a vscode extension for improved ucode support
within the IDE. It features auto complete suggestions as well as
semantic syntax highlighting, compile time error reporting and support
for JSDoc type annotations in variable or function declarations.

As part of the vscode extensions, a new ucode library "uscope" is
introduced which provides facilities for static ucode source analysis,
e.g. to determine the validity scope of variable declarations or to
properly tokenize ucode sources.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>