nl80211: properly support split_wiphy_dump for single phys
When performing a `NL80211_CMD_GET_WIPHY` request for a single wiphy with
`split_wiphy_dump` set to true, the expectation is that only a single phy
object is returned, like when performing the same request without splitting,
yet the current implementation returns a sparse array with only the
requested phy index populated instead.

Fix this issue by special-casing NL80211_CMD_GET_WIPHY command requests
with a set `split_wiphy_dump` attribute. Also implicitly set the
`NLM_F_DUMP` flag in this case as it is required for the kernel to send
all information.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>