socket: properly handle async `connect(2)` errors in socket.connect()
Connected `O_NONBLOCK` sockets in failure state are reported as `POLLOUT`
by `poll(2)` as well, so use `getsockopt(SO_ERROR)` to query for potential
errors.

Skip those sockets that have errors and continue to give other connect
candidates a chance to succeed but remember the error in order to be able
to report it if all attempts fail.

Fixes: #302
Signed-off-by: Jo-Philipp Wich <jo@mein.io>