vm: implement PVAL opcode
Introduce a new opcode PVAL (peek value) which is similar to LVAL.
The difference between LVAL and PVAL is that the latter leaves the
receiver value on stack while the former consumes it.

This behaviour is needed for implementing proper short circuiting optional
chaining behavior later on without having to copy and swapping stack values
around.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>