compiler, vm: rework optional chaining and function call semantics
Introduce two new opcodes `PVAL` (peek value) and `JMPNT` (jump if not type)
and utilize them in the compiler to implement proper short-circuiting
behavior for optional chaining operators `?.`, `?.[`, and `?.(`.

The `PVAL` opcode takes a receiver and a property name value from the stack,
pops the property name value, and pushes the result of `receiver[property]`
back onto the stack. It is very similar to `LVAL` but leaves the receiver
in place instead of popping it as well.

The `JMPNT` opcode checks the type of the value at a certain stack depth
and, if the type does not match, pops all stack slots from the checked
value to the top of the stack, pushes a single `null`, and then jumps to
the destination address. When the type matches, execution continues after
the opcode sequence. The opcode is followed by an unsigned 32-bit argument
in the format DDD|TTTTTTTTTTTTT|AAAAAAAAAAAAAAAA, with bits 0-15 encoding
the signed relative jump address (-0x7fff to +0x7fff), bits 16-28 being a
bitmap of types to check for, and bits 29-31 (0x00 to 0x07) specifying the
stack depth at which to check the value.

Also increase the bytecode version due to incompatible opcode changes.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>