vm: adjust JMPNT opcode behavior
Always check the value at the top of the stack but pop `depth` times
additional values on type mismatch. This fixes optionally chained
method calls (`foo.bar?.()`) where we need to pop two stack items on
type mismatch but check the function value at the top of the stack.

For `?.` and `?.[` optional chaining operations, JMPNT is called with
a depth of `0` so the behavior remains unchanged for these cases.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>