ucode: Add 'operator execute'
It can be useful to call a resource as if it was a function. This commit adds a way to do so.

When a non-callable is supplied to uc_vm_insn_call(), and it is an object or has a prototype, then the object/prototype stack is searched for a callable named 'operator()', and if found, that is called instead.

So in the script
var(...); is a functional synonym to var['operator()']();

Signed-off-by: Isaac de Wolff <idewolff@vincitech.nl>