struct: align `X` and `Z` length & truncation semantics with `*`
Improve the usability of `X` and `Z` formats with variable length inputs;
default to unpacking the entire buffer remainder and packing the entire
input if no specific width is given in the format string. When a width is
specified, then unpack/encode and pack/decode that many raw bytes at most.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>