socket: fix `recv()` incorrectly reporting unrelated errors
When `recv()` is invoked without length argument, `ucv_to_integer()` is not
called and `errno` not reset. This may cause the subsequent check to
incorrectly report a receive error when an unrelated operation failed
since `errno` was last cleared.

Rework the code to only check `errno` if an actual length argument is
present.

Fixes: #332
Signed-off-by: Jo-Philipp Wich <jo@mein.io>